1008378: added sample for binding MS SQL with Dapper using UrlAdaptor.u

+++ Gantt/IntegrationPage/Binding Mysql DB using Custom Adaptor/Blazor Web App/Gantt_MySQL/Data/TaskDataModel.cs
@@ -8,7 +8,6 @@
         public DateTime? EndDate { get; set; }
         public int? ParentId { get; set; }
         public string? Duration { get; set; }    
-        public decimal Progress { get; set; }            
-
+        public int Progress { get; set; }
     }
 }
